Continuous Integration!! Using travisCI
Every time a push or merge request is made, a build will be sent to travis.
A link to the build requests should be available somewhere from the github
repository:

https://guthub.com/ajjl/compiler

Click on commits, or go to the branch that was pushed.
It might be a good idea to establish some sort of workflow where commits are
only pushed to master after they have been proven to not break the build
(or cause an increase in test failures. Once we get some tests set up).
This may not work when we move to arm only stuff, but it should be good for now.

git-svn-id: https://svn.divms.uiowa.edu/repos/cs4980/groupc@110 a9721779-b9a4-4edf-a86f-66882cd6156f