Updates Makefile. Adds lexsupport.o target. Gets rid of error output for make clean.
The "make clean" command was producing a false error if the files it was trying
to clean weren't present. This has been resolved by adding the "-f" option to
the rm commands in the clean target.

The lexsupport.o target was added to build the new stuff that jones just made.

git-svn-id: https://svn.divms.uiowa.edu/repos/cs4980/groupc@156 a9721779-b9a4-4edf-a86f-66882cd6156f