Adds infrastructure for google-test framework. Add -Winline compiler flag.
Adds infrastructure for the google test framework. We may or may not end up
using this.

Also adds compiler flag -Winline
Inline functions usually offer the same performacne benefits of macros, but are
less prone to bugs. If the compiler is unable to inline a function, it will give
a warning.

git-svn-id: https://svn.divms.uiowa.edu/repos/cs4980/groupc@174 a9721779-b9a4-4edf-a86f-66882cd6156f