Updated README with cmake instructions
CMake is weird, but worth it because i dont know how to link to testing framworks
with only make. We should probably still have an old fashioned Makefile that works.
CMake should allow us to more easily have tests.
If your not used to cmake, its probably easier to figure it out if I show you,

git-svn-id: https://svn.divms.uiowa.edu/repos/cs4980/groupc@182 a9721779-b9a4-4edf-a86f-66882cd6156f