Updates Makefile
environment.h uses some c++11 features. Because of this it needs to be explicitly
called within the makefile to ensure that it is compiled with the "-std=c++11"
flag. This causes a precompiled header to be rpoduced, and the .gitignore and the
```make clean``` targets were updated to be aware of these types of files.

git-svn-id: https://svn.divms.uiowa.edu/repos/cs4980/groupc@942 a9721779-b9a4-4edf-a86f-66882cd6156f